Agent/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymysql==1.0.2
python-dotenv==1.0.0
numpy==1.24.3
tensorflow==2.10.0
pillow==9.5.0
keras==2.10.0